fungsi get_pagination_data di pisah ke DataTableTrait
