+++ app/Http/Livewire/Table/DataTableTrait.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace App\Http\Livewire\Table;
+
+
+trait DataTableTrait {
+    
+    public function get_pagination_data ()
+    {
+        switch ($this->name) {
+            case 'user':
+                $users = $this->model::search($this->search)
+                    ->orderBy($this->sortField, $this->sortAsc ? 'asc' : 'desc')
+                    ->paginate($this->perPage);
+
+                return [
+                    "view" => 'livewire.table.user',
+                    "users" => $users,
+                    "data" => array_to_object([
+                        'href' => [
+                            'create_new' => route('user.new'),
+                            'create_new_text' => 'Buat User Baru',
+                            'export' => '#',
+                            'export_text' => 'Export'
+                        ]
+                    ])
+                ];
+                break;
+
+            default:
+                # code...
+                break;
+        }
+    }
+}
