+++ app/Http/Livewire/Table/Main.php
@@ -7,7 +7,7 @@ use Livewire\WithPagination;
 
 class Main extends Component
 {
-    use WithPagination;
+    use WithPagination, DataTableTrait;
 
     public $model;
     public $name;
@@ -28,34 +28,6 @@ class Main extends Component
         }
 
         $this->sortField = $field;
-    }
-
-    public function get_pagination_data ()
-    {
-        switch ($this->name) {
-            case 'user':
-                $users = $this->model::search($this->search)
-                    ->orderBy($this->sortField, $this->sortAsc ? 'asc' : 'desc')
-                    ->paginate($this->perPage);
-
-                return [
-                    "view" => 'livewire.table.user',
-                    "users" => $users,
-                    "data" => array_to_object([
-                        'href' => [
-                            'create_new' => route('user.new'),
-                            'create_new_text' => 'Buat User Baru',
-                            'export' => '#',
-                            'export_text' => 'Export'
-                        ]
-                    ])
-                ];
-                break;
-
-            default:
-                # code...
-                break;
-        }
     }
 
     public function delete_item ($id)
